requirements.txt
torch>=2.2.0
torchvision>=0.17.0
numpy>=1.26.0
Pillow>=10.2.0
albumentations>=1.4.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
tqdm>=4.66.0
tensorboard>=2.12.0    # Para visualización del entrenamiento
pytest>=8.0.0          # Para testing
pytest-cov>=4.1.0      # Para cobertura de código 